Move backoffice backend DI out of config folder
