+++ src/apps/backoffice/backend/dependency-injection/application_production.yaml
