+++ src/apps/backoffice/backend/dependency-injection/application_staging.yaml
