+++ src/apps/backoffice/backend/dependency-injection/application_test.yaml
@@ -3,9 +3,9 @@ imports:
 
 services:
   Backoffice.Backend.EnvironmentArranger:
-    class: ../../../../../../tests/Contexts/Shared/infrastructure/mongo/MongoEnvironmentArranger
+    class: ../../../../../tests/Contexts/Shared/infrastructure/mongo/MongoEnvironmentArranger
     arguments: ['@Shared.MongoConnectionManager']
 
   Backoffice.Backend.ElasticEnvironmentArranger:
-    class: ../../../../../../tests/Contexts/Shared/infrastructure/elastic/ElasticEnvironmentArranger
+    class: ../../../../../tests/Contexts/Shared/infrastructure/elastic/ElasticEnvironmentArranger
     arguments: ['@Shared.ElasticConnectionManager']
