+++ src/apps/backoffice/backend/dependency-injection/apps/application.yaml
@@ -1,18 +1,18 @@
 services:
   Apps.Backoffice.Backend.MongoConfig:
     factory:
-      class: ../../../../../../Contexts/Backoffice/Courses/infrastructure/persistence/BackofficeMongoConfigFactory
+      class: ../../../../../Contexts/Backoffice/Courses/infrastructure/persistence/BackofficeMongoConfigFactory
       method: 'createConfig'
 
   Apps.Backoffice.Backend.ElasticConfig:
     factory:
-      class: ../../../../../../Contexts/Backoffice/Courses/infrastructure/persistence/BackofficeElasticConfigFactory
+      class: ../../../../../Contexts/Backoffice/Courses/infrastructure/persistence/BackofficeElasticConfigFactory
       method: 'createConfig'
 
   Apps.Backoffice.Backend.controllers.StatusGetController:
-    class: ../../../controllers/StatusGetController
+    class: ../../controllers/StatusGetController
     arguments: []
 
   Apps.Backoffice.Backend.controllers.CoursesGetController:
-    class: ../../../controllers/CoursesGetController
+    class: ../../controllers/CoursesGetController
     arguments: ['@Shared.QueryBus']
