+++ src/apps/backoffice/backend/routes/courses.route.ts
@@ -1,8 +1,10 @@
 import { Express } from 'express';
-import container from '../config/dependency-injection';
+import container from '../dependency-injection';
 import { CoursesGetController } from '../controllers/CoursesGetController';
 
 export const register = (app: Express) => {
-  const coursesGetController: CoursesGetController = container.get('Apps.Backoffice.Backend.controllers.CoursesGetController');
+  const coursesGetController: CoursesGetController = container.get(
+    'Apps.Backoffice.Backend.controllers.CoursesGetController'
+  );
   app.get('/courses', coursesGetController.run.bind(coursesGetController));
 };
