+++ src/apps/backoffice/backend/routes/status.route.ts
@@ -1,5 +1,5 @@
 import { Express } from 'express';
-import container from '../config/dependency-injection';
+import container from '../dependency-injection';
 import StatusController from '../controllers/StatusGetController';
 
 export const register = (app: Express) => {
